Update `membership` action
# Support `hazelcast` username

In GitHub, usernames and organisations are unique.

The `membership` action is often triggered with `github.event.pull_request.head.repo.owner.login` - so for a _non_-fork PR the value will be `hazelcast` (the owner of a repo).
`hazelcast` _isn't_ a member of the `hazelcast` organisation - it _is_ the organisation - so the check should pass and the PR should run.
Updated check to suit.


# Avoid deprecated set-output in membership action

https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Avoids [warning](https://github.com/hazelcast/hazelcast-python-client/actions/runs/13280122985):
> Check membership of given user
The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/