Update `membership` action to support `hazelcast` username (#69)
In GitHub, usernames and organisations are unique.

The `membership` action is often triggered with
`github.event.pull_request.head.repo.owner.login` - so for a _non_-fork
PR the value will be `hazelcast` (the owner of a repo). `hazelcast`
_isn't_ a member of the `hazelcast` organisation - it _is_ the
organisation - so the check should pass and the PR should run. Updated
check to suit.